Refactor
Change list:
    Move jnc.foreign.{=>spi.}ForeignProvider
    Remove method jnc.foreign.Foreign.provider
    Cleaner.register Return a runnable
    Rename LayoutBuilder.{add=>new}Field

+++ jnc-runtime/src/test/java/jnc/foreign/internal/SleepUtil.java
@@ -29,7 +29,7 @@ public interface SleepUtil {
         System.gc();
         boolean interrupted = false;
         try {
-            int sleep = 1;
+            long sleep = 1;
             for (int i = 0; i < 12 && condition.getAsBoolean(); ++i) {
                 try {
                     TimeUnit.MILLISECONDS.sleep(sleep);
